+++ custom_components/visual_dashboard_editor/const.py
@@ -1,9 +1,9 @@
 """Constants for Visual Dashboard Editor."""
 
 DOMAIN = "visual_dashboard_editor"
-VERSION = "0.2.18"
+VERSION = "0.2.19"
 
-PANEL_ELEMENT = "visual-dashboard-editor-panel-v15"
+PANEL_ELEMENT = "visual-dashboard-editor-panel-v16"
 PANEL_TITLE = "Visual Dashboard Editor"
 PANEL_ICON = "mdi:vector-square-edit"
 PANEL_URL = "visual-dashboard-editor"